refact: move PcglFooter to PCGL component folder
