+++ src/js/components/Pcgl/PcglFooter.tsx
