+++ src/js/components/Util/DefaultLayout.tsx
@@ -6,7 +6,7 @@ import SiteHeader from '@/components/SiteHeader';
 import SiteSider from '@/components/SiteSider';
 import SiteFooter from '@/components/SiteFooter';
 import PageHeader from '@/components/PageHeader';
-import PcglFooter from '@/components/PcglFooter';
+import PcglFooter from '@/components/Pcgl/PcglFooter';
 import ScopedTitle from '@/components/Scope/ScopedTitle';
 import { useSelectedScope } from '@/features/metadata/hooks';
 import { useIsInCatalogueMode, useSidebarMenuItems } from '@/hooks/navigation';
@@ -41,8 +41,7 @@ const DefaultLayout = () => {
           <Content>
             <Outlet />
           </Content>
-          <SiteFooter />
-          {PCGL_MODE ? <PcglFooter /> : null}
+          {PCGL_MODE ? <PcglFooter /> : <SiteFooter />}
           {/* Overview has its own way of rendering a back-to-top button, so we only render this if we're not on the overview page: */}
           {page !== BentoRoute.Overview ? (
             <FloatButton.BackTop className="float-btn-pos" target={() => document.getElementById('content-layout')!} />
